move Message into models
