+++ cibo/actions/commands/finalize.py
@@ -7,9 +7,9 @@ from peewee import IntegrityError
 from cibo.actions.__action__ import Action
 from cibo.client import Client
 from cibo.exception import ClientIsLoggedIn, PlayerAlreadyExists, PlayerNotRegistered
-from cibo.messages.__message__ import Message
 from cibo.models.data.item import Item
 from cibo.models.data.player import Player
+from cibo.models.message import Message
 
 
 class Finalize(Action):
