+++ cibo/actions/commands/inventory.py
@@ -5,7 +5,7 @@ from typing import List
 from cibo.actions.__action__ import Action
 from cibo.client import Client
 from cibo.exception import ClientNotLoggedIn
-from cibo.messages.__message__ import Message
+from cibo.models.message import Message
 
 
 class Inventory(Action):
