+++ cibo/actions/commands/quit.py
@@ -6,7 +6,7 @@ from typing import List, Optional, Tuple
 from cibo.actions.__action__ import Action
 from cibo.client import Client
 from cibo.exception import ClientIsLoggedIn
-from cibo.messages.__message__ import Message
+from cibo.models.message import Message
 
 
 class Quit(Action):
