+++ cibo/actions/commands/register.py
@@ -7,8 +7,8 @@ from marshmallow import ValidationError
 from cibo.actions.__action__ import Action
 from cibo.client import Client
 from cibo.exception import ClientIsLoggedIn, PlayerAlreadyExists, PlayerNotFound
-from cibo.messages.__message__ import Message
 from cibo.models.data.player import Player, PlayerSchema
+from cibo.models.message import Message
 
 
 class Register(Action):
