+++ cibo/actions/commands/say.py
@@ -5,7 +5,7 @@ from typing import List, Tuple
 from cibo.actions.__action__ import Action
 from cibo.client import Client
 from cibo.exception import ActionMissingArguments, ClientNotLoggedIn
-from cibo.messages.__message__ import Message, MessageRoute
+from cibo.models.message import Message, MessageRoute
 
 
 class Say(Action):
