+++ cibo/actions/disconnect.py
@@ -4,7 +4,7 @@ from typing import List, Optional
 
 from cibo.actions.__action__ import Action
 from cibo.client import Client
-from cibo.messages.__message__ import Message
+from cibo.models.message import Message
 
 
 class Disconnect(Action):
